rework falloc functionality
The previous FrameAllocator functionality was quite limited. Which is to say, it lacked support for reference counting frames—which by nature puts a stronger emphasis on the caller to be responsible with acquired frames. Being unable to reference count also made shared mappings challenging, such as in the case of shared libraries.

Hopefully this new approach (with ref counting, frame borrowing, and distinct frame locking) more sophisticated use of the system's page frames.